Split year picker grid view
